# more test improvements