… src/test/java/io/vertx/amqp/SenderTypeTest.java