Some more changes to use a vertx-proton test server to more closely test specific client behaviours of interest, rather than doing send+recieve through a full broker. The 'bonus speedup' benefit for these particular classes are...

Before:
Tests run: 17, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 30.457 sec - in io.vertx.amqp.SenderTypeTest
Tests run: 20, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 14.624 sec - in io.vertx.amqp.ReceptionTypeTest

After:
Tests run: 18, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.387 sec - in io.vertx.amqp.SenderTypeTest
Tests run: 22, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.764 sec - in io.vertx.amqp.ReceptionTypeTest


Changes follow on from #45, with more improvement still remaining possible for further classes.