@gemmellr 2020-08-25
Cherry picked to 3.9 branch in cbd095d24a31c74c18bb197d2fff058569c0a9ce